requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
numpy==1.24.3
aiosqlite==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
websockets==12.0
requests==2.31.0
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
asyncio-mqtt==0.16.1
aiofiles==23.2.1
bcrypt==4.1.2
PyJWT==2.8.0